+++ src/main/java/com/perimeterx/api/PerimeterX.java
@@ -27,7 +27,10 @@ package com.perimeterx.api;
 
 import com.perimeterx.api.activities.ActivityHandler;
 import com.perimeterx.api.activities.BufferedActivityHandler;
-import com.perimeterx.api.providers.*;
+import com.perimeterx.api.providers.CombinedIPProvider;
+import com.perimeterx.api.providers.DefaultHostnameProvider;
+import com.perimeterx.api.providers.HostnameProvider;
+import com.perimeterx.api.providers.IPProvider;
 import com.perimeterx.api.proxy.DefaultReverseProxy;
 import com.perimeterx.api.proxy.ReverseProxy;
 import com.perimeterx.api.remoteconfigurations.DefaultRemoteConfigManager;
@@ -44,7 +47,6 @@ import com.perimeterx.models.activities.UpdateReason;
 import com.perimeterx.models.configuration.PXConfiguration;
 import com.perimeterx.models.configuration.PXDynamicConfiguration;
 import com.perimeterx.models.exceptions.PXException;
-import com.perimeterx.models.risk.CustomParameters;
 import com.perimeterx.models.risk.PassReason;
 import com.perimeterx.utils.PXLogger;
 
@@ -58,6 +60,7 @@ import java.net.URISyntaxException;
  * <p>
  * Created by shikloshi on 03/07/2016.
  */
+
 public class PerimeterX {
 
     private static final PXLogger logger = PXLogger.getLogger(PerimeterX.class);
@@ -72,12 +75,6 @@ public class PerimeterX {
     private ReverseProxy reverseProxy;
 
     private void init(PXConfiguration configuration) throws PXException {
-        if (configuration.isDebugMode()) {
-            PXLogger.setDebugLevel();
-        } else {
-            PXLogger.setErrorLevel();
-        }
-
         logger.debug(PXLogger.LogReason.DEBUG_INITIALIZING_MODULE);
         this.configuration = configuration;
         hostnameProvider = new DefaultHostnameProvider();
@@ -179,7 +176,6 @@ public class PerimeterX {
         logger.debug(PXLogger.LogReason.DEBUG_COOKIE_MISSING);
         if (serverValidator.verify(context)) {
             logger.debug(PXLogger.LogReason.DEBUG_COOKIE_VERSION_FOUND, context.getCookieVersion());
-            return;
         }
     }