+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## [v5.4.0](https://github.com/PerimeterX/perimeterx-java-sdk/compare/v1.0.17...HEAD) (2018-12-13)
+- Removed logback log implementation.
+- Removed debugMode configuration, instead use log level configuration per logger implementation.
+
 ## [v5.3.0](https://github.com/PerimeterX/perimeterx-java-sdk/compare/v1.0.17...HEAD) (2018-12-12)
 - Fixed http components memory leak ([HTTPASYNC-116](https://issues.apache.org/jira/browse/HTTPASYNC-116))
 - Added custom params to async activities (page_requested, block)